requirements.txt
supabase==1.0.3
pypdf==3.15.4
python-dotenv==1.0.0
numpy==1.24.3
redis==4.5.4
pillow==9.5.0
torch==2.0.0
transformers==4.28.1
requests==2.31.0